requirements.txt
flask>=3.0
pypdf>=4.0
werkzeug>=3.0
python-docx>=1.0
openpyxl>=3.1
python-pptx>=0.6
Pillow>=10.0